+++ sdk/core/azure_core/src/http/pager.rs
@@ -4,12 +4,20 @@
 use crate::http::{headers::HeaderName, response::Response};
 use async_trait::async_trait;
 use futures::{stream::unfold, FutureExt, Stream};
-use std::{fmt, future::Future, pin::Pin, task};
+#[cfg(not(target_arch = "wasm32"))]
+use std::str::FromStr;
+use std::{
+    fmt,
+    future::Future,
+    pin::Pin,
+    sync::{Arc, Mutex},
+    task,
+};
 use typespec::Error;
 use typespec_client_core::http::{DeserializeWith, Format, JsonFormat};
 
 /// The result of fetching a single page from a [`Pager`], whether there are more pages or paging is done.
-pub enum PagerResult<P, N> {
+pub enum PagerResult<P, N: AsRef<str>> {
     /// There are more pages the [`Pager`] may fetch using the `next` token.
     More { response: P, next: N },
     /// The [`Pager`] is done and there are no additional pages to fetch.
@@ -29,7 +37,7 @@ impl<P, F> PagerResult<Response<P, F>, String> {
     }
 }
 
-impl<P, N: fmt::Debug> fmt::Debug for PagerResult<P, N> {
+impl<P, N: fmt::Debug + AsRef<str>> fmt::Debug for PagerResult<P, N> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::More { next, .. } => f
@@ -202,16 +210,16 @@ impl<P: Page> ItemIterator<P> {
     /// ```
     pub fn from_callback<
         // This is a bit gnarly, but the only thing that differs between the WASM/non-WASM configs is the presence of Send bounds.
-        #[cfg(not(target_arch = "wasm32"))] N: Send + 'static,
+        #[cfg(not(target_arch = "wasm32"))] N: AsRef<str> + Send + 'static,
         #[cfg(not(target_arch = "wasm32"))] F: Fn(Option<N>) -> Fut + Send + 'static,
         #[cfg(not(target_arch = "wasm32"))] Fut: Future<Output = Result<PagerResult<P, N>, typespec::Error>> + Send + 'static,
-        #[cfg(target_arch = "wasm32")] N: 'static,
+        #[cfg(target_arch = "wasm32")] N: AsRef<str> + 'static,
         #[cfg(target_arch = "wasm32")] F: Fn(Option<N>) -> Fut + 'static,
         #[cfg(target_arch = "wasm32")] Fut: Future<Output = Result<PagerResult<P, N>, typespec::Error>> + 'static,
     >(
         make_request: F,
     ) -> Self {
-        Self::from_stream(iter_from_callback(make_request))
+        Self::from_stream(iter_from_callback(make_request, || None, |_| {}))
     }
 
     /// Creates a [`ItemIterator<P>`] from a raw stream of [`Result<P>`](typespec::Result<P>) values.
@@ -238,6 +246,7 @@ impl<P: Page> ItemIterator<P> {
     pub fn into_pages(self) -> PageIterator<P> {
         PageIterator {
             stream: self.stream,
+            next: Default::default(),
         }
     }
 }
@@ -289,6 +298,7 @@ impl<P: Page> fmt::Debug for ItemIterator<P> {
 pub struct PageIterator<P> {
     #[pin]
     stream: Pin<BoxedStream<P>>,
+    next: Arc<Mutex<Option<String>>>,
 }
 
 impl<P> PageIterator<P> {
@@ -384,16 +394,44 @@ impl<P> PageIterator<P> {
     /// ```
     pub fn from_callback<
         // This is a bit gnarly, but the only thing that differs between the WASM/non-WASM configs is the presence of Send bounds.
-        #[cfg(not(target_arch = "wasm32"))] N: Send + 'static,
+        #[cfg(not(target_arch = "wasm32"))] N: AsRef<str> + FromStr + Send + 'static,
         #[cfg(not(target_arch = "wasm32"))] F: Fn(Option<N>) -> Fut + Send + 'static,
         #[cfg(not(target_arch = "wasm32"))] Fut: Future<Output = Result<PagerResult<P, N>, typespec::Error>> + Send + 'static,
-        #[cfg(target_arch = "wasm32")] N: 'static,
+        #[cfg(target_arch = "wasm32")] N: AsRef<str> + FromStr + 'static,
         #[cfg(target_arch = "wasm32")] F: Fn(Option<N>) -> Fut + 'static,
         #[cfg(target_arch = "wasm32")] Fut: Future<Output = Result<PagerResult<P, N>, typespec::Error>> + 'static,
     >(
         make_request: F,
-    ) -> Self {
-        Self::from_stream(iter_from_callback(make_request))
+    ) -> Self
+    where
+        <N as FromStr>::Err: fmt::Debug,
+    {
+        let next = Arc::new(Mutex::new(None::<String>));
+
+        let get_clone = next.clone();
+        let set_clone = next.clone();
+        let stream = iter_from_callback(
+            make_request,
+            move || {
+                if let Ok(next_guard) = get_clone.lock() {
+                    return next_guard
+                        .clone()
+                        .map(|n| n.parse().expect("valid next link"));
+                }
+
+                None
+            },
+            move |next_token| {
+                if let Ok(mut next_guard) = set_clone.lock() {
+                    *next_guard = next_token.map(Into::into);
+                }
+            },
+        );
+
+        Self {
+            stream: Box::pin(stream),
+            next,
+        }
     }
 
     /// Creates a [`PageIterator<P>`] from a raw stream of [`Result<P>`](typespec::Result<P>) values.
@@ -408,7 +446,58 @@ impl<P> PageIterator<P> {
     ) -> Self {
         Self {
             stream: Box::pin(stream),
+            next: Default::default(),
         }
+    }
+
+    /// Advance the `PageIterator` to the page referenced by `next_link`.
+    ///
+    /// You should call this before iterating the [`Stream`] or results may be unpredictable.
+    ///
+    /// # Examples
+    ///
+    /// Using a result of a call to [`PageIterator::next_link`] in another process, you can create a new `PageIterator`
+    /// that, when first iterated, will get the next page of results.
+    ///
+    /// ``` no_run
+    /// use azure_identity::DefaultAzureCredential;
+    /// use azure_security_keyvault_secrets::SecretClient;
+    /// use futures::stream::TryStreamExt as _;
+    ///
+    /// # #[tokio::main]
+    /// # async fn main() -> azure_core::Result<()> {
+    /// let client = SecretClient::new("https://my-vault.vault.azure.net", DefaultAzureCredential::new()?, None)?;
+    /// let mut pager = client.list_secret_properties(None)?
+    ///     .into_pages()
+    ///     .with_next_link("next_link_from_another_pager".to_string());
+    ///
+    /// while let Some(secrets) = pager.try_next().await? {
+    ///     let secrets = secrets.into_body().await?;
+    ///     for secret in secrets.value {
+    ///         println!("{:?}", secret.id);
+    ///     }
+    /// }
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn with_next_link(self, next_link: String) -> Self {
+        if let Ok(mut next_guard) = self.next.lock() {
+            *next_guard = Some(next_link);
+        }
+
+        self
+    }
+
+    /// Gets the next link or continuation token for the current page.
+    ///
+    /// Pass this to [`PageIterator::with_next_link`] to create a `PageIterator` that, when first iterated,
+    /// will return the next page. You can use this to page results across separate processes.
+    pub fn next_link(&self) -> Option<String> {
+        if let Ok(next) = self.next.lock() {
+            return next.clone();
+        }
+
+        None
     }
 }
 
@@ -439,35 +528,53 @@ enum State<T> {
 fn iter_from_callback<
     P,
     // This is a bit gnarly, but the only thing that differs between the WASM/non-WASM configs is the presence of Send bounds.
-    #[cfg(not(target_arch = "wasm32"))] N: Send + 'static,
+    #[cfg(not(target_arch = "wasm32"))] N: AsRef<str> + Send + 'static,
     #[cfg(not(target_arch = "wasm32"))] F: Fn(Option<N>) -> Fut + Send + 'static,
     #[cfg(not(target_arch = "wasm32"))] Fut: Future<Output = Result<PagerResult<P, N>, typespec::Error>> + Send + 'static,
-    #[cfg(target_arch = "wasm32")] N: 'static,
+    #[cfg(not(target_arch = "wasm32"))] G: Fn() -> Option<N> + Send + 'static,
+    #[cfg(not(target_arch = "wasm32"))] S: Fn(Option<&str>) + Send + 'static,
+    #[cfg(target_arch = "wasm32")] N: AsRef<str> + 'static,
     #[cfg(target_arch = "wasm32")] F: Fn(Option<N>) -> Fut + 'static,
     #[cfg(target_arch = "wasm32")] Fut: Future<Output = Result<PagerResult<P, N>, typespec::Error>> + 'static,
+    #[cfg(target_arch = "wasm32")] G: Fn() -> Option<N> + 'static,
+    #[cfg(target_arch = "wasm32")] S: Fn(Option<&str>) + 'static,
 >(
     make_request: F,
+    get_next: G,
+    set_next: S,
 ) -> impl Stream<Item = Result<P, Error>> + 'static {
     unfold(
-        // We flow the `make_request` callback through the state value so that we can avoid cloning.
-        (State::Init, make_request),
-        |(state, make_request)| async move {
+        // We flow the `make_request` callback, 'get_next', and `set_next` through the state value so that we can avoid cloning.
+        (State::Init, make_request, get_next, set_next),
+        |(mut state, make_request, get_next, set_next)| async move {
+            if let Some(next_link) = get_next() {
+                state = State::More(next_link);
+            }
             let result = match state {
                 State::Init => make_request(None).await,
-                State::More(c) => make_request(Some(c)).await,
-                State::Done => return None,
+                State::More(n) => make_request(Some(n)).await,
+                State::Done => {
+                    set_next(None);
+                    return None;
+                }
             };
             let (item, next_state) = match result {
-                Err(e) => return Some((Err(e), (State::Done, make_request))),
+                Err(e) => return Some((Err(e), (State::Done, make_request, get_next, set_next))),
                 Ok(PagerResult::More {
                     response,
-                    next: continuation,
-                }) => (Ok(response), State::More(continuation)),
-                Ok(PagerResult::Done { response }) => (Ok(response), State::Done),
+                    next: next_token,
+                }) => {
+                    set_next(Some(next_token.as_ref()));
+                    (Ok(response), State::More(next_token))
+                }
+                Ok(PagerResult::Done { response }) => {
+                    set_next(None);
+                    (Ok(response), State::Done)
+                }
             };
 
-            // Flow 'make_request' through to avoid cloning
-            Some((item, (next_state, make_request)))
+            // Flow 'make_request', 'get_next', and 'set_next' through to avoid cloning
+            Some((item, (next_state, make_request, get_next, set_next)))
         },
     )
 }
@@ -476,7 +583,7 @@ fn iter_from_callback<
 mod tests {
     use crate::http::{
         headers::{HeaderName, HeaderValue},
-        Pager, PagerResult, RawResponse, StatusCode,
+        PageIterator, Pager, PagerResult, RawResponse, Response, StatusCode,
     };
     use async_trait::async_trait;
     use futures::{StreamExt as _, TryStreamExt as _};
@@ -486,7 +593,7 @@ mod tests {
     #[derive(Deserialize, Debug, PartialEq, Eq)]
     struct Page {
         pub items: Vec<i32>,
-        pub page: i32,
+        pub page: Option<i32>,
     }
 
     #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
@@ -596,7 +703,7 @@ mod tests {
                 "page-1".to_string(),
                 Page {
                     items: vec![1],
-                    page: 1
+                    page: Some(1)
                 }
             ),
             pages[0].as_ref().unwrap()
@@ -605,5 +712,110 @@ mod tests {
         let err = pages[1].as_ref().unwrap_err();
         assert_eq!(&typespec::error::ErrorKind::Other, err.kind());
         assert_eq!("yon request didst fail", format!("{}", err));
+    }
+
+    #[tokio::test]
+    async fn page_iterator_with_next_link_and_tracking() {
+        let make_callback = || {
+            |continuation: Option<String>| async move {
+                match continuation.as_deref() {
+                    None => Ok(PagerResult::More {
+                        response: RawResponse::from_bytes(
+                            StatusCode::Ok,
+                            Default::default(),
+                            r#"{"items":[1],"page":1}"#,
+                        )
+                        .into(),
+                        next: "next-token-1".to_string(),
+                    }),
+                    Some("next-token-1") => Ok(PagerResult::More {
+                        response: RawResponse::from_bytes(
+                            StatusCode::Ok,
+                            HashMap::from([(
+                                HeaderName::from_static("x-test-header"),
+                                HeaderValue::from_static("page-2"),
+                            )])
+                            .into(),
+                            r#"{"items":[2],"page":2}"#,
+                        )
+                        .into(),
+                        next: "next-token-2".to_string(),
+                    }),
+                    Some("next-token-2") => Ok(PagerResult::Done {
+                        response: RawResponse::from_bytes(
+                            StatusCode::Ok,
+                            HashMap::from([(
+                                HeaderName::from_static("x-test-header"),
+                                HeaderValue::from_static("page-3"),
+                            )])
+                            .into(),
+                            r#"{"items":[3]}"#,
+                        )
+                        .into(),
+                    }),
+                    _ => {
+                        panic!("Unexpected continuation value: {:?}", continuation)
+                    }
+                }
+            }
+        };
+
+        // Create the first PageIterator.
+        let mut first_pager: PageIterator<Response<Page>> =
+            PageIterator::from_callback(make_callback());
+
+        // Should start with no next_link.
+        assert_eq!(first_pager.next_link(), None);
+
+        // Advance to the first page.
+        let first_page = first_pager
+            .next()
+            .await
+            .expect("expected first page")
+            .expect("expected successful first page")
+            .into_body()
+            .await
+            .expect("expected page");
+        assert_eq!(first_page.page, Some(1));
+        assert_eq!(first_page.items, vec![1]);
+
+        // next_link should point to second page.
+        let next_link = first_pager
+            .next_link()
+            .expect("expected next_link from first page");
+        assert_eq!(next_link, "next-token-1");
+
+        // Create the second PageIterator.
+        let mut second_pager: PageIterator<Response<Page>> =
+            PageIterator::from_callback(make_callback()).with_next_link(next_link);
+
+        // Should start with link to second page.
+        assert_eq!(second_pager.next_link(), Some("next-token-1".into()));
+
+        // Advance to second page.
+        let second_page = second_pager
+            .next()
+            .await
+            .expect("expected second page")
+            .expect("expected successful second page")
+            .into_body()
+            .await
+            .expect("expected page");
+        assert_eq!(second_page.page, Some(2));
+        assert_eq!(second_page.items, vec![2]);
+        assert_eq!(second_pager.next_link(), Some("next-token-2".into()));
+
+        // Advance to last page.
+        let last_page = second_pager
+            .next()
+            .await
+            .expect("expected last page")
+            .expect("expected successful last page")
+            .into_body()
+            .await
+            .expect("expected page");
+        assert_eq!(last_page.page, None);
+        assert_eq!(last_page.items, vec![3]);
+        assert_eq!(second_pager.next_link(), None);
     }
 }